Extract connect-oauth pure helpers into a tested route-adjacent module
Move the pure helper clusters (stored config parsing, config merging,
session/exchange state, query parsing, PKCE) from connect-oauth.tsx into
connect-oauth-config.ts, keeping the route as the UI orchestrator. Add
characterization tests for previously untested helpers, including an
RFC 7636 S256 PKCE vector. Behavior-preserving move for #543.
