+++ packages/worker/client/routes/connect-oauth.node.test.ts
@@ -1,13 +1,24 @@
+import { utf8ToBase64Url } from '@kody-internal/shared/base64.ts'
 import { expect, test } from 'vitest'
 import {
+	createCodeChallenge,
+	createCodeVerifier,
+	decodeBase64Payload,
+	formatMissingSetupFields,
 	formatOAuthExchangeFailure,
+	isMostlyPrintable,
 	isOAuthExchangeSessionExpired,
+	isSafeExternalUrl,
 	mergeConnectOauthConfig,
 	parseConnectOauthNextSteps,
+	parseExtraParams,
+	parseOptionalUrl,
+	parseProviderSetupInstructions,
+	parseScopes,
 	parseSessionConnectOauthConfig,
 	parseStoredIntegrationConfig,
 	summarizeStoredSetupState,
-} from './connect-oauth.tsx'
+} from './connect-oauth-config.ts'
 
 test('connect OAuth helpers parse stored integrations, merge reconnect configs, and derive provider defaults', () => {
 	const parsed = parseStoredIntegrationConfig(
@@ -764,4 +775,136 @@ test('parseConnectOauthNextSteps accepts suggestion payload and drops unsafe URL
 	})
 	expect(parseConnectOauthNextSteps(null)).toBeNull()
 	expect(parseConnectOauthNextSteps({ guidance: 'x' })).toBeNull()
+	expect(
+		parseConnectOauthNextSteps({
+			guidance: 'g',
+			integrationName: 'x',
+			suggestions: 'not-an-array',
+			createHelpersCta: { label: 'a', prompt: 'b' },
+		}),
+	).toBeNull()
+	expect(
+		parseConnectOauthNextSteps({
+			guidance: 'g',
+			integrationName: 'x',
+			suggestions: [{ listingId: 1 }],
+			createHelpersCta: { label: 'a', prompt: 'b' },
+		}),
+	).toEqual({
+		guidance: 'g',
+		integrationName: 'x',
+		suggestions: [],
+		createHelpersCta: { label: 'a', prompt: 'b' },
+	})
+	expect(
+		parseConnectOauthNextSteps({
+			guidance: 'g',
+			integrationName: 'x',
+			suggestions: [],
+			createHelpersCta: { label: 1, prompt: 'b' },
+		}),
+	).toBeNull()
+})
+
+test('formatMissingSetupFields enumerates missing credential fields', () => {
+	expect(formatMissingSetupFields([])).toBe('Ready to connect.')
+	expect(formatMissingSetupFields(['client ID'])).toBe(
+		'Enter your client ID to continue.',
+	)
+	expect(formatMissingSetupFields(['client ID', 'client secret'])).toBe(
+		'Enter your client ID and client secret to continue.',
+	)
+	expect(
+		formatMissingSetupFields(['client ID', 'client secret', 'redirect URI']),
+	).toBe('Enter your client ID, client secret and redirect URI to continue.')
+})
+
+test('isSafeExternalUrl and parseOptionalUrl accept only usable http(s) URLs', () => {
+	expect(isSafeExternalUrl('https://example.com/path')).toBe(true)
+	expect(isSafeExternalUrl('http://localhost:8787')).toBe(true)
+	expect(isSafeExternalUrl('javascript:alert(1)')).toBe(false)
+	expect(isSafeExternalUrl('ftp://files.example.com')).toBe(false)
+	expect(isSafeExternalUrl('not a url')).toBe(false)
+
+	expect(parseOptionalUrl(null)).toBeNull()
+	expect(parseOptionalUrl('')).toBeNull()
+	expect(parseOptionalUrl('https://example.com/a?b=1')).toBe(
+		'https://example.com/a?b=1',
+	)
+	expect(parseOptionalUrl('not a url')).toBeNull()
+})
+
+test('decodeBase64Payload and isMostlyPrintable characterize base64 instruction decoding', () => {
+	expect(decodeBase64Payload('!!!')).toBeNull()
+	expect(
+		decodeBase64Payload(utf8ToBase64Url('Open the developer console.')),
+	).toBe('Open the developer console.')
+	expect(decodeBase64Payload('not!valid')).toBeNull()
+
+	expect(isMostlyPrintable('')).toBe(false)
+	expect(isMostlyPrintable('printable text\nwith tabs\tand returns\r')).toBe(
+		true,
+	)
+	expect(isMostlyPrintable('\u0001\u0002\u0003\u0004\u0005')).toBe(false)
+	// Threshold is strictly greater than 0.85 printable characters.
+	expect(isMostlyPrintable(`${'ok'.repeat(9)}\u0001\u0002`)).toBe(true)
+	expect(isMostlyPrintable('ok\u0001\u0002')).toBe(false)
+})
+
+test('parseExtraParams and parseScopes tolerate query-string shapes', () => {
+	expect(parseExtraParams(null)).toEqual({})
+	expect(parseExtraParams('')).toEqual({})
+	expect(
+		parseExtraParams('{"prompt":"consent","access_type":"offline"}'),
+	).toEqual({
+		prompt: 'consent',
+		access_type: 'offline',
+	})
+	expect(parseExtraParams('{"n":1,"b":true,"x":null}')).toEqual({
+		n: '1',
+		b: 'true',
+		x: 'null',
+	})
+	expect(parseExtraParams('["not","object"]')).toEqual({})
+	expect(parseExtraParams('{')).toEqual({})
+
+	expect(parseScopes(null)).toEqual([])
+	expect(parseScopes('')).toEqual([])
+	expect(parseScopes('   ')).toEqual([])
+	expect(parseScopes('repo, read:user  openid')).toEqual([
+		'repo',
+		'read:user',
+		'openid',
+	])
+	expect(parseScopes('["repo","read:user"]')).toEqual(['repo', 'read:user'])
+	expect(parseScopes('[not-json')).toEqual(['[not-json'])
+	expect(parseScopes('{"not":"array"}')).toEqual(['{"not":"array"}'])
+})
+
+test('parseProviderSetupInstructions decodes base64 payloads when printable', () => {
+	const plain = 'Visit the provider console and create an OAuth app.'
+	const encoded = utf8ToBase64Url(plain)
+	expect(parseProviderSetupInstructions(null)).toBeNull()
+	expect(parseProviderSetupInstructions('')).toBeNull()
+	expect(parseProviderSetupInstructions('   ')).toBeNull()
+	expect(parseProviderSetupInstructions(plain)).toBe(plain)
+	expect(parseProviderSetupInstructions(`base64:${encoded}`)).toBe(plain)
+	expect(parseProviderSetupInstructions(encoded)).toBe(plain)
+	expect(parseProviderSetupInstructions('base64:!!!')).toBe('base64:!!!')
+})
+
+test('PKCE helpers match RFC 7636 S256 vector and current verifier shape', async () => {
+	const rfcVerifier = 'dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk'
+	await expect(createCodeChallenge(rfcVerifier)).resolves.toBe(
+		'E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM',
+	)
+
+	const verifier = createCodeVerifier()
+	// Current implementation samples 64 random bytes → 86-char base64url (not the
+	// 43-char / 32-byte length recommended by RFC 7636 appendix B).
+	expect(verifier).toMatch(/^[A-Za-z0-9_-]+$/)
+	expect(verifier).toHaveLength(86)
+	await expect(createCodeChallenge(verifier)).resolves.toMatch(
+		/^[A-Za-z0-9_-]+$/,
+	)
 })
